+++ CHANGELOG.md
@@ -3,6 +3,12 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 
+## [1.2.1] - 2023-02-23
+### Added
+- IInteraction.GameObject();
+### Changed
+- IInteraction.cs param name gameObject to invoker
+
 ## [1.2.0] - 2023-01-23
 ### Added
 - A bunch of new scripts